Tests/LibWeb/Text/expected/css/shadow-root-host-keyframes-invalidation.txt

opacity before keyframes: 1
opacity after keyframes insert: 0.25
opacity after keyframes delete: 1
